feat(dram): update dram configuration and initialization
- Change dram size type from uint64_t to uint32_t across multiple files
- Update board-specific dram parameters and add new initialization functions
- Implement power management and voltage control for dram
- Add new register definitions for system control
- Link libdram.a in CMake configuration

+++ include/drivers/sys-dram.h
@@ -106,7 +106,7 @@ typedef struct {
  * 
  * @return The size of the DRAM in bytes.
  */
-uint64_t sunxi_get_dram_size();
+uint32_t sunxi_get_dram_size();
 
 /**
  * @brief Initialize the DRAM.
